+++ hutool-core/src/main/java/cn/hutool/core/codec/Base64.java
@@ -299,7 +299,7 @@ public class Base64 {
 	 * base64解码
 	 *
 	 * @param base64 被解码的base64字符串
-	 * @return 被加密后的字符串
+	 * @return 解码后的bytes
 	 */
 	public static byte[] decode(CharSequence base64) {
 		return Base64Decoder.decode(base64);